requirements.txt
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
PyPDF2==3.0.1
python-dotenv==1.1.1
reportlab==4.4.2
Werkzeug==3.1.3
WTForms==3.2.1
